pyproject.toml
[project]
name = "eyeamme"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "bcrypt>=5.0.0",
    "boto3>=1.42.19",
    "botocore>=1.42.19",
    "cryptography>=46.0.3",
    "fastapi>=0.128.0",
    "gunicorn>=23.0.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
]
[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.uv.build-backend]
module-root = "src"
